--- tests/robottelo/test_api_utils.py
@@ -1,7 +1,0 @@
-"""Unit tests for :mod:`robottelo.api.utils`."""
-from robottelo.api import utils
-
-
-def test_one_to_one_names():
-    """Test :func:`robottelo.api.utils.one_to_one_names`."""
-    assert utils.one_to_one_names('person') == {'person_name', 'person_id'}
